paper_trading/requirements.txt
# Python 3.10+
discord.py>=2.3.2
python-dotenv>=1.0.1
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
aiohttp>=3.9.0
apscheduler>=3.10.4
tzdata>=2024.1
